use url params

+++ controllers/IUserController.go
@@ -3,6 +3,6 @@ package controllers
 import "go-exercise/dtos"
 
 type IUserController interface {
-	GetUser(id int64) dtos.UserDTO
-	UpdateUser(id int64, name string, email string, avatar string)
+	GetUser(id int) dtos.UserDTO
+	UpdateUser(id int, name string, email string, avatar string)
 }
