+++ controllers/UserController.go
@@ -9,7 +9,7 @@ type UserController struct {
 	repositories.IUserRepository
 }
 
-func (controller *UserController) GetUser(id int64) dtos.UserDTO {
+func (controller *UserController) GetUser(id int) dtos.UserDTO {
 	user, err := controller.FindById(id)
 	if err != nil {
 		panic(err)
@@ -17,7 +17,7 @@ func (controller *UserController) GetUser(id int64) dtos.UserDTO {
 	return dtos.UserDTO{user.Id, user.Name, user.Email, user.Avatar}
 }
 
-func (controller *UserController) UpdateUser(id int64, name string, email string, avatar string) {
+func (controller *UserController) UpdateUser(id int, name string, email string, avatar string) {
 	user, err := controller.FindById(id)
 	if err != nil {
 		panic(err)
