+++ dtos/UserDTO.go
@@ -1,7 +1,7 @@
 package dtos
 
 type UserDTO struct {
-	Id     int64
+	Id     int
 	Name   string
 	Email  string
 	Avatar string
