+++ repositories/IUserRepository.go
@@ -3,6 +3,6 @@ package repositories
 import "go-exercise/entities"
 
 type IUserRepository interface {
-	FindById(id int64) (entities.User, error)
+	FindById(id int) (entities.User, error)
 	Save(user *entities.User) error
 }
